requirements.txt
autocommand==2.2.2
bcrypt==4.0.1
cffi==1.15.1
cheroot==9.0.0
CherryPy==18.8.0
cherrypy-jinja2==2.2
cryptography==40.0.2
inflect==6.0.4
jaraco.collections==4.1.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==9.1.0
paramiko==3.1.0
portend==3.1.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
PyNaCl==1.5.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sshtunnel==0.4.0
tempora==5.2.2
typing_extensions==4.5.0
zc.lockfile==3.0.post1
